patch out intentionally debounced dispatchEvent, refactor renderer scripts
