+++ src/extra/css/fixtures.css
@@ -1,14 +1,3 @@
-/* osake specific */
-@media (prefers-color-scheme: light) {
-	.osake-icon.adaptive {
-		color: #1f1f1f;
-	}
-}
-
-.osake-icon {
-	color: white;
-}
-
 /* hide apple music's logo */
 .logo {
 	display: none !important;
